Toggle input

A binary toggle Shiny input.

sa_toggle(
  inputId,
  label,
  value = FALSE,
  true = "Yes",
  false = "No",
  color = "blue",
  text_color = "black",
  width = NULL
)

update_sa_toggle(
  session,
  inputId,
  label = NULL,
  value = NULL,
  true = NULL,
  false = NULL
)

Arguments

inputId
The id of the input object.

label
The label to set for the input object.

value
logical, default toggle state.

true
character, left label for TRUE state.

false
character, right label for FALSE state.

color
active state color.

text_color
color of left and right value label text.

width
The width of the input, e.g. '400px', or '100 shiny::validateCssUnit.

session
The session object passed to function given to shinyServer.

Value

HTML element that can be added to a UI definition.

Details

This input returns TRUE or FALSE similar to shiny::checkboxInput.

Examples

if (interactive()) {

ui <- function(request) {
fluidPage(
  use_access(),
  fluidRow(bookmarkButton()),
  fluidRow(
    column(3,
      actionButton("btn1", "Set to TRUE"),
      actionButton("btn2", "Change label"),
      sa_toggle(
        "toggle1", "Show text A", FALSE, "On", "Off",
        "firebrick", "white", 200
      ),
      sa_toggle(
        "toggle2", "Show text B", TRUE, "On", "Off",
        "yellowgreen", "black", "100%"
      )
    ),
    column(9, textOutput("txt1"), textOutput("txt2"))
  )
)
}

server <- function(input, output, session) {
  output$txt <- renderText(paste("Toggle input is set to", input$toggle1))
  observeEvent(input$btn1, {
    update_sa_toggle(session, "toggle1", value = TRUE)
  })
  observeEvent(input$btn2, {
    update_sa_toggle(session, "toggle1", label = "Show text A 2",
                     true = "Yes", false = "No")
  })
}

shinyApp(ui, server, enableBookmarking = "url")

}